moved inputfield to attendee folder + chat icon placement fix
