--- src/components/atoms/InputField/InputField.module.scss
@@ -1,15 +1,0 @@
-.inputFieldWrapper {
-  position: relative;
-}
-
-.iconEnd {
-  position: absolute;
-  top: 0;
-  right: 0;
-  height: 100%;
-  display: flex;
-
-  svg {
-    margin: auto 0;
-  }
-}
